Create De Martino Expanded theory
This series of codes compile our efforts to expand the theory introduced by De Martino et al to gather the transmission probability for a graphene sheet under a perpendicular magnetic field. We expanded the theory by developing a method to test the wavefunctions solutions in each region of the boundary problem, calculating the transmission amplitude and the transmission probability, not only reproducing his theory but also expanding it to make it valid for a monolayer graphene transistor with magnetic field on the channel. We added all the realistic parameters from our graphene chips, for example, fermi velocity, lentgh, width, voltage, magnetic field, etc. Other parameters can be verified in the code.