Add files via upload
Code for plotting the transmission polar plots from PRL 98, 066802 (2007). Beware that in line (*transmission amplitude*), the complex conjugate of cosine must be calculated previously by hand and then inserted in the code. For some reason, if mathematica calculates it by itself, it get the wrong numerical values and does not plot the right transmission plots.